+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * [154](https://github.com/mlebreuil/netbox-contract/issues/154) Fix edit and delete bulk operations on dimensions and invoice lines.
 * [153](https://github.com/mlebreuil/netbox-contract/issues/153) Enforce uniquness of accounting dimensions.
 * Adds a status ( Active or Inactive ) to accounitng dimensions.
+* [151](https://github.com/mlebreuil/netbox-contract/issues/151) Fix accounting line and dimensions search.
 
 ### Version 2.2.1